+++ Delta.CertXplorer/src/Delta.CertXplorer/CertManager/Wrappers/ObjectWrapper.cs
@@ -15,6 +15,7 @@ namespace Delta.CertXplorer.CertManager.Wrappers
             if (item == null) return null;
             if (item is Certificate) return new CapiCertificateWrapper((Certificate)item);
             if (item is CertificateRevocationList) return new CapiCrlWrapper((CertificateRevocationList)item);
+            if (item is CertificateTrustList) return new CapiCtlWrapper((CertificateTrustList)item);
             if (item is X509Certificate2) return new X509CertificateWrapper2((X509Certificate2)item);
             if (item is X509Certificate) return new X509CertificateWrapper((X509Certificate)item);
             return item;